requirements.txt
pyyaml>=6.0
tqdm>=4.66.0
streamlit>=1.28